move arithmetic tests from period.test_ops
